Refactoring of directives, fix of multiples char suffixes, start removing ambiguity of LPAREN/'('
Suffixes of int/float with multiples chars such as "UL" are now correctly produced.
Function like macro and callback have been partially added.
LPAREN is currently causing an issue with '(' because token_list doesn't contains it, adding it would resolve '(' not preceded by whitespace but
will also create an ambiguity between define directives so this should be fixed to allow a correct production of token list in any situations.