Fix representation of nested structure.
Nested structure should be probably handled differently to ease further generation of code based
on AST. Currently Nested struct are declared as a singular object and this is conflicting with specifier_qualifier_list
visit as all others specifiers and qualifiers are strings only.

Special attention should be paid to solve this in an elegant manner.