Add first draft of AST transformer to ctypes/python like typedefs.
Ctypes and Python generator should allow at least generation of enumeration as Python object.
Each generator is using Python format strings as templating engine in purpose to create dynamically classes.
Struct and unions are in work in progress state but it shouldn't not be so hard to achieve as well.